requirements.txt

flask==3.0.0
flask-cors==4.0.0
paho-mqtt==1.6.1
